// src/app/models/ArticuloDTO.ts
export class ArticuloDTO {
  id: number;
  nombre: string;
  abreviatura: string;
  codigoArt: string;
  stockMin: number;
  stockMax: number;
  costo: number;
  precio: number;
  proveedorId: number;
  unidadMedidaId: number;
  marcaId: number;
  rubroId: number;
  subRubroId: number;
  habilitacion: number;
}